@@ -14,7 +14,7 @@ exports.newId = () => {
 exports.concatBuffer = (buf_1, buf_2) => {
   const len_1 = buf_1.length;
   const len_2 = buf_2.length;
-  const buf = new Buffer(len_1 + len_2);
+  const buf = Buffer.alloc(len_1 + len_2);
   buf_1.copy(buf, 0, 0, len_1);
   buf_2.copy(buf, len_1, 0, len_2);
   return buf;
