@@ -7,20 +7,18 @@ const Invocation = require('./invocation');
 exports.requestEncode = (id, data, options) => {
   const req = new Request(id);
   const arr = data.serverSignature.split(':');
-  let serviceName = data.serverSignature;
-  let version = '1.0.0';
+  const attachments = Object.assign({
+    dubbo: '5.3.0',
+    path: data.serverSignature,
+  }, data.requestProps);
   if (arr.length === 2) {
-    serviceName = arr[0];
-    version = arr[1];
+    attachments.path = arr[0];
+    attachments.version = arr[1];
   }
   req.data = new Invocation({
     methodName: data.methodName,
     args: data.args,
-    attachments: Object.assign({
-      dubbo: '5.3.0',
-      path: serviceName,
-      version,
-    }, data.requestProps),
+    attachments,
   });
   return req.encode(options);
 };
