@@ -69,7 +69,7 @@ class Request {
       const inv = this.data;
       output.writeUTF(inv.attachments[Constants.DUBBO_VERSION_KEY] || DUBBO_VERSION);
       output.writeUTF(inv.attachments[Constants.PATH_KEY]);
-      output.writeUTF(inv.attachments[Constants.VERSION_KEY]);
+      output.writeUTF(inv.attachments[Constants.VERSION_KEY] || '');
 
       output.writeUTF(inv.methodName);
       output.writeUTF(utils.getJavaArgsDesc(inv.args));
@@ -109,8 +109,8 @@ class Request {
         const attachments = {
           [Constants.DUBBO_VERSION_KEY]: field,
           [Constants.PATH_KEY]: input.readUTF(),
-          [Constants.VERSION_KEY]: input.readUTF(),
         };
+        const version = input.readUTF();
         const methodName = input.readUTF();
         const desc = input.readUTF();
         const methodArgSigs = utils.desc2classArray(desc);
@@ -122,7 +122,7 @@ class Request {
         }
         Object.assign(attachments, input.readObject());
 
-        const serverSignature = attachments.version ? `${attachments.path}:${attachments.version}` : attachments.path;
+        const serverSignature = version ? attachments.path + ':' + version : attachments.path;
         data = {
           methodName,
           serverSignature,
